Use the right parameters to configure datasource
This patch uses the user and password instead of setting basic auth that
is used as credentials when connecting to the Grafana server. In our
case we are using the grafana configuration on the minion to connect to
Grafana. This patch also adds an option to set the datasource as the
default one.